+++ package.json
@@ -25,9 +25,11 @@
     "dotenv": "^16.0.0",
     "ejs": "^3.1.6",
     "express": "^4.17.2",
+    "jsonwebtoken": "^8.5.1",
     "mongodb": "^4.3.1",
     "mysql": "^2.18.1",
     "socket.io": "^4.4.1",
+    "uuid": "^8.3.2",
     "uuidv4": "^6.2.12",
     "xss": "^1.0.10"
   }
